+++ src/main/java/forestry/core/commands/SpeciesArgument.java
@@ -1,10 +1,5 @@
 package forestry.core.commands;
 
-import com.google.common.collect.ImmutableList;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Random;
 import java.util.concurrent.CompletableFuture;
 
 import net.minecraft.commands.SharedSuggestionProvider;
@@ -21,36 +16,7 @@ import com.mojang.brigadier.suggestion.SuggestionsBuilder;
 import forestry.api.genetics.ISpecies;
 import forestry.api.genetics.ISpeciesType;
 
-public class SpeciesArgument implements ISpeciesArgumentType<ISpecies<?>> {
-	private final ISpeciesType<?, ?> type;
-	private final Collection<String> examples;
-
-	public SpeciesArgument(ISpeciesType<?, ?> type) {
-		this.type = type;
-
-		// Generate 3 random examples
-		this.examples = new ArrayList<>(3);
-
-		Random rand = new Random();
-		ImmutableList<ResourceLocation> ids = type.getAllSpeciesIds().asList();
-
-		// if there are fewer than 3 registered species, don't loop forever
-		int numExamples = Math.min(3, ids.size());
-
-		while (this.examples.size() < numExamples) {
-			String example = ids.get(rand.nextInt(ids.size())).toString();
-
-			if (!this.examples.contains(example)) {
-				this.examples.add(example);
-			}
-		}
-	}
-
-	@Override
-	public ISpeciesType<?, ?> getType() {
-		return this.type;
-	}
-
+public record SpeciesArgument(ISpeciesType<?, ?> type) implements ISpeciesArgumentType<ISpecies<?>> {
 	@Override
 	public ISpecies<?> parse(final StringReader reader) throws CommandSyntaxException {
 		ResourceLocation location = ResourceLocation.read(reader);
@@ -67,10 +33,5 @@ public class SpeciesArgument implements ISpeciesArgumentType<ISpecies<?>> {
 	@Override
 	public <SOURCE> CompletableFuture<Suggestions> listSuggestions(final CommandContext<SOURCE> context, final SuggestionsBuilder builder) {
 		return SharedSuggestionProvider.suggestResource(this.type.getAllSpeciesIds(), builder);
-	}
-
-	@Override
-	public Collection<String> getExamples() {
-		return this.examples;
 	}
 }