+++ src/main/java/forestry/lepidopterology/commands/CommandButterfly.java
@@ -20,16 +20,21 @@ import com.mojang.brigadier.builder.ArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 
+import forestry.api.lepidopterology.genetics.IButterflySpeciesType;
 import forestry.core.commands.CommandHelpers;
 import forestry.core.commands.GiveSpeciesCommand;
+import forestry.core.commands.ModifyGenomeCommand;
 import forestry.core.utils.SpeciesUtil;
 import forestry.lepidopterology.features.LepidopterologyEntities;
 
 public class CommandButterfly {
 	public static ArgumentBuilder<CommandSourceStack, ?> register() {
+		IButterflySpeciesType type = SpeciesUtil.BUTTERFLY_TYPE.get();
+
 		return Commands.literal("butterfly")
 				.then(CommandButterflyKill.register())
-				.then(GiveSpeciesCommand.register(SpeciesUtil.BUTTERFLY_TYPE.get()));
+				.then(GiveSpeciesCommand.register(type))
+				.then(ModifyGenomeCommand.register(type));
 	}
 
 	public static class CommandButterflyKill {